include config command in ALL 'Learn about deployment...
Hi,
Following azure documentation can be tricky sometimes and I have been finding myself looping through the rabbit hole of blogs to accomplish one task. Just knowing whether you are in ASM or ARM was a struggle for myself and I am assuming others now that this message is at the beginning of a lot of blog posts. Whichever deployment model the blog is covering, the `azure config mode` asm or arm needs to be included in this message, IMO. It's beyond frustrating to start going through the blog to find out you are in the wrong deployment model and to find out whatever resource was spun up is basically wrong. Better to start the average developer off on the right step. It would be a huge help!